+++ bht/errors.py
@@ -8,14 +8,23 @@ class BhtError(Exception):
 
 class BhtPipelineError(BhtError):
     """Pipeline error"""
+
     pass
 
 
 class BhtPathError(BhtError):
     """File or Dir path error"""
+
     pass
 
 
 class BhtResultError(BhtError):
     """Result error"""
+
+    pass
+
+
+class BhtCsvError(BhtError):
+    """Result error"""
+
     pass